Revert "Add completion for jobs and variable files."
