+++ commands/internal/flaghelpers/job_flag.go
@@ -2,13 +2,9 @@ package flaghelpers
 
 import (
 	"errors"
-	"fmt"
 	"strings"
 
 	"github.com/concourse/go-concourse/concourse"
-	"github.com/jessevdk/go-flags"
-
-	"github.com/concourse/fly/rc"
 )
 
 type JobFlag struct {
@@ -35,48 +31,4 @@ func (job *JobFlag) UnmarshalFlag(value string) error {
 	job.JobName = vs[1]
 
 	return nil
-}
-
-func (flag *JobFlag) Complete(match string) []flags.Completion {
-	fly := parseFlags()
-
-	target, err := rc.LoadTarget(fly.Target)
-	if err != nil {
-		return []flags.Completion{}
-	}
-
-	err = target.Validate()
-	if err != nil {
-		return []flags.Completion{}
-	}
-
-	team := target.Team()
-	comps := []flags.Completion{}
-	vs := strings.SplitN(match, "/", 2)
-
-	if len(vs) == 1 {
-		pipelines, err := team.ListPipelines()
-		if err != nil {
-			return comps
-		}
-
-		for _, pipeline := range pipelines {
-			if strings.HasPrefix(pipeline.Name, vs[0]) {
-				comps = append(comps, flags.Completion{Item: pipeline.Name + "/"})
-			}
-		}
-	} else if len(vs) == 2 {
-		jobs, err := team.ListJobs(vs[0])
-		if err != nil {
-			return comps
-		}
-
-		for _, job := range jobs {
-			if strings.HasPrefix(job.Name, vs[1]) {
-				comps = append(comps, flags.Completion{Item: fmt.Sprintf("%s/%s", vs[0], job.Name)})
-			}
-		}
-	}
-
-	return comps
 }
