+++ commands/internal/flaghelpers/pipeline_flag.go
@@ -1,6 +1,7 @@
 package flaghelpers
 
 import (
+	"os"
 	"strings"
 
 	"github.com/jessevdk/go-flags"
@@ -8,10 +9,27 @@ import (
 	"github.com/concourse/fly/rc"
 )
 
+type flyCommand struct {
+	Target rc.TargetName `short:"t" long:"target" description:"Concourse target name"`
+}
+
+var fly flyCommand
+
 type PipelineFlag string
 
 func (flag *PipelineFlag) Complete(match string) []flags.Completion {
-	fly := parseFlags()
+	// Prevent go-flags from recursing
+	goFlagsCompletion, hasCompletion := os.LookupEnv("GO_FLAGS_COMPLETION")
+	os.Unsetenv("GO_FLAGS_COMPLETION")
+	defer func() {
+		if hasCompletion {
+			os.Setenv("GO_FLAGS_COMPLETION", goFlagsCompletion)
+		}
+	}()
+
+	parser := flags.NewParser(&fly, flags.HelpFlag|flags.PassDoubleDash)
+	parser.NamespaceDelimiter = "-"
+	parser.Parse()
 
 	target, err := rc.LoadTarget(fly.Target)
 	if err != nil {
